Fix MR comments

+++ gr-a3sat/lib/conv_dec_impl.h
@@ -2,8 +2,6 @@
 #define INCLUDED_A3SAT_CONV_DEC_IMPL_H
 
 #include <a3sat/conv_dec.h>
-#include <unordered_map>
-#include <map>
 
 namespace gr {
     namespace a3sat {
@@ -78,7 +76,7 @@ namespace gr {
              * @param output_items
              * @return noutput_items
              */
-            int general_work(int noutput_items,
+            int general_work(uint16_t noutput_items,
                              gr_vector_int &ninput_items,
                              gr_vector_const_void_star &input_items,
                              gr_vector_void_star &output_items);
@@ -89,7 +87,7 @@ namespace gr {
              * @param noutput_items
              * @param ninput_items_required
              */
-            void forecast(int noutput_items, gr_vector_int &ninput_items_required);
+            void forecast(uint16_t noutput_items, gr_vector_int &ninput_items_required);
 
             /*!
              * @brief Calculates the “distance” between the transmitted and the received symbols.
